Optini

Optini.AbstractAlgorithm

Optini.BacktrackingLineSearch

Optini.BacktrackingLineSearch

Optini.Bisection

Optini.ExactLineSearch

Optini.ExactLineSearch

Optini.Fibonacci

Optini.Fibonacci

Optini.GoldenSection

Optini.GradientDescent

Optini.GradientDescent

Optini.GradientDescent

Optini.InterpolationLineSearch

Optini.InterpolationLineSearch

Optini.LineSearch

Optini.LineSearch

Optini.MultipleIdentity

Optini.MultipleIdentity

Optini.MultivariateAlgorithm

Optini.Newton

Optini.PreviousDecreaseInitial

Optini.PreviousDecreaseInitial

Optini.QuadraticFit

Optini.QuadraticFit

Optini.QuadraticInitial

Optini.QuadraticInitial

Optini.Solution

Optini.StaticInitial

Optini.StaticInitial

Optini.StaticLineSearch

Optini.Steepest

Optini.StrongWolfeLineSearch

Optini.StrongWolfeLineSearch

Optini.TrustRegion

Optini.TrustRegion

Optini.UnivariateAlgorithm

Optini.bracket

Optini.optimize

Optini.AbstractAlgorithm — Type

AbstractAlgorithm

Abstract supertype of all optimization algorithms.

Optini.BacktrackingLineSearch — Type

BacktrackingLineSearch{C, T}

BacktrackingLineSearch starts from an initial step length guess and gradually backtracks until the Armijo sufficient decrease condition is met.

Fields

c::C: the constant for the Armijo condition in the open interval (0, 1)

scale::T: the scale factor to shrink the step length

Optini.BacktrackingLineSearch — Method

BacktrackingLineSearch(; kwargs...)

Initiate BacktrackingLineSearch.

Keywords

c=1e-4: the constant for the Armijo condition in the open interval (0, 1)

scale=0.5: the scale factor to shrink the step length

Optini.Bisection — Type

Bisection <: UnivariateAlgorithm

Bisection search is a root-finding algorithm by maintainining a bracket of opposing derivative signs. At each iteration the algorithm shrinks the bracket by half until a critical point is located or the bracket is sufficiently small.

Optini.ExactLineSearch — Type

ExactLineSearch{A<:UnivariateAlgorithm, O1, O2}

ExactLineSearch computes the optimal step length using univariate methods. Since most univariate algorithms require repeated calculations of the objective to shrink the bracket around a local minimum, ExactLineSearch is expected to be expensive.

Fields

alg::A: the univariate algorithm

bracket_options::O1: named tuple of bracketing options

alg_options::O2: named tuple of the univariate algorithm options

Optini.ExactLineSearch — Method

ExactLineSearch(; kwargs...)

Initiate a ExactLineSearch. Use keyword arguments to tweak the bracketing phase and the univariate algorithm.

Keywords

alg::UnivariateAlgorithm: the univariate algorithm to optimize the step length

step=0.01: bracketing step size

scale=2.0: bracketing scale factor

bracket_maxiter::Integer=100: bracketing maximum number of iterations

reltol=1e-6: relative tolerance

abstol=1e-12: absolute tolerance

maxiter::Integer=1_000: maximum number of iterations

Optini.Fibonacci — Type

Fibonacci{T} <: UnivariateAlgorithm

The Fibonacci search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum. The ratio is computed from the Fibonacci sequence.

Optini.Fibonacci — Type

Fibonacci([ϵ=0.01])

Initialize Fibonacci search algorithm with optional argument ϵ in the interval (0, 1) to control how close the last iterate is to the previous one. Refer to Kochenderfer and Wheeler's "Algorithms for Optimization" for more information.

Optini.GoldenSection — Type

GoldenSection <: UnivariateAlgorithm

The golden section search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum with the golden ratio as an approximation to the Fibonacci sequence.

Optini.GradientDescent — Type

GradientDescent{I, M}

Aliasing LineSearch with steepest descent.

Optini.GradientDescent — Method

GradientDescent(initial, method)

Initiate GradientDescent.

Arguments

initial: starts line search with an initial step length guess

method: computes the appropriate next step

Optini.GradientDescent — Method

GradientDescent(; kwargs...)

Initiate GradientDescent.

Keywords

initial=StaticInitial(0.001): starts line search with an initial step length guess

method=StaticLineSearch(): computes the appropriate next step

Optini.InterpolationLineSearch — Type

InterpolationLineSearch{C, T}

InterpolationLineSearch uses quadratic and cubic interpolations to compute the step length satisfying the Armijo sufficient decrease condition.

Fields

c::C: the constant for the Armijo condition in the open interval (0, 1)

ϵ::T: the minimum step length decrease from the initial step length

Optini.InterpolationLineSearch — Method

InterpolationLineSearch(; kwargs...)

Initiate InterpolationLineSearch.

Keywords

c=1e-4: the constant for the Armijo condition in the open interval (0, 1)

ϵ=1e-6: the minimum step length decrease from the initial step length

Optini.LineSearch — Type

LineSearch{D, I, M} <: MultivariateAlgorithm

LineSearch optimizes the objective function by first choosing a direction, then search along this direction for an appropriate step.

Fields

direction::D: determines the search direction

initial::I: the method to guess an initial step length

method::: the method to compute the appropriate next step

Optini.LineSearch — Method

LineSearch(; kwargs...)

Initiate LineSearch algorithm.

Keywords

direction=Steepest(): determines the search direction

initial=StaticInitial(1.0): the method to guess an initial step length

method=BacktrackingLineSearch(): the method to compute the appropriate next step

Optini.MultipleIdentity — Type

MultipleIdentity{B, S}

Modify the Hessian by adding some multiples, τ, of the identity matrix, as described in Nocedal and Wright's Numerical Optimization.

Fields

β::B: constant to compute appropriate τ

scale::S: scale factor to expand τ

Optini.MultipleIdentity — Method

MultipleIdentity(; beta=1e-3, scale=2.0)

Initiate MultipleIdentity method to modify the Hessian.

Optini.MultivariateAlgorithm — Type

MultivariateAlgorithm

Abstract supertype of multivariate optimization algorithms.

Optini.Newton — Type

Newton{M}

Newton dispatches to the Newton direction for LineSearch algorithm and uses the Hessian in TrustRegion models.

Fields

modify::M: the Hessian modification method to enforce positive definiteness

Optini.PreviousDecreaseInitial — Type

PreviousDecreaseInitial{T, D<:Ref} <: AbstractInitial{T}

PreviousDecreaseInitial assumes the first-order decrease is the same as the previous iteration:

$α₀ = αₖ₋₁\frac{∇f^\intercal_{k-1} p_{k-1}}{∇f^\intercal_k p_k}$

Fields

α::T: default step length for the first iteration

prev_decrease::D: save the last iteration's first-order decrease

Optini.PreviousDecreaseInitial — Method

PreviousDecreaseInitial([α::T=1.0]) where {T}

Initialize PreviousDecreaseInitial, with prev_decrease defaulting to Ref(T(NaN)).

Optini.QuadraticFit — Type

QuadraticFit{T} <: UnivariateAlgorithm

The quadratic fit algorithm constructs a quadratic function from three points at each iteration as an approximation to the objective. The analytical solution to the quadratic model is examined as the new minimizer candidate to update the three points.

Optini.QuadraticFit — Type

QuadraticFit([ϵ=0.01])

Initialize quadratic fit search algorithm with optional argument ϵ in the interval (0, 1) to choose a midpoint as mid = ϵ * lower + (1 - ϵ) * upper.

Optini.QuadraticInitial — Type

QuadraticInitial{T} <: AbstractInitial{T}

QuadraticInitial assumes $ϕ(α₀) - ϕ(0) = fₖ - fₖ₋₁$ and uses quadratic interpolation with $ϕ(α₀)$, $ϕ(0)$, $ϕ′(0)$ to compute the initial guess.

Fields

α::T: default step length for the first iteration

prev_f::Ref{T}: save the last iteration's function value

Optini.QuadraticInitial — Method

QuadraticInitial([α::T=1.0]) where {T}

Initialize QuadraticInitial, with prev_f defaulting to Ref(T(NaN)).

Optini.Solution — Type

Solution{Tx, Tf}

Solution type for optimization models.

Optini.StaticInitial — Type

StaticInitial{T} <: AbstractInitial{T}

Static initial α₀ for line search methods. Newton and quasi-Newton algorithms for instance should use α₀ = 1 as the initial guess.

Optini.StaticInitial — Type

StaticInitial([α=1.0])

Initialize StaticInitial{T}.

Optini.StaticLineSearch — Type

StaticLineSearch

StaticLineSearch directly uses the initial guess as the new step length.

Optini.Steepest — Type

Steepest

Steepest dispatches to the opposite of the gradient direction for LineSearch.

Optini.StrongWolfeLineSearch — Type

StrongWolfeLineSearch{C1, C2, S, T}

StrongWolfeLineSearch computes a step length that satifies both the Armijo sufficient decrease and the curvature conditions.

Fields

c₁::C1: Armijo condition constant

c₂::C2: curvature condition constant

scale::S: scale factor to expand step length bracket

α_max::T: maximum step length

Optini.StrongWolfeLineSearch — Method

StrongWolfeLineSearch(; kwargs...)

Initiate StrongWolfeLineSearch.

Keywords

c1=1e-4: Armijo condition constant

`c2=0.9: curvature condition constant

scale=2.0: scale factor to expand step length bracket

alphamax=100_000: maximum step length

Optini.TrustRegion — Type

TrustRegion{H, M, D, R<:Ref, E} <: MultivariateAlgorithm

TrustRegion builds a quadratic model within a radius from the current point to approximate the objective function and uses the model to choose the next step.

Fields

hessian::H: the method to compute or approximate the Hessian

method::M<:AbstractTrustRegionMethod: the method to compute the appropriate next step

Δ₀::D: the initial radius

Δ::R: a Ref containing the current radius

η::E: the threshold to control when TrustRegion will take a step

Optini.TrustRegion — Method

TrustRegion(; kwargs...)

Initiate TrustRegion algorithm.

Keywords

hessian=Newton(): the method to compute or approximate the Hessian

method=TwoDimSubspace(): the method to compute the appropriate next step

delta=5.0: the initial radius

deltamax=100.0: the maximum radius

eta=0.1: the threshold to control when TrustRegion will take a step

Optini.UnivariateAlgorithm — Type

UnivariateAlgorithm

Abstract supertype of univariate optimization algorithms.

Optini.bracket — Method

bracket(f::Function, x::T=0.0; kwargs...) where {T<:AbstractFloat}

Create a bracket around a local minimum from the intial point x. Return the bracket in a tuple if successful, and nothing otherwise.

Keywords

step=0.01: the step size to expand the bracket

scale=2.0: the scale factor for step size at each iteration

maxiter::Integer=100: the maximum number of iterations

Optini.optimize — Method

optimize(f::Function, lower::Real, upper::Real; kwargs...)

Compute the minimizer of the function f using the bracket (lower, upper).

Keywords

alg::UnivariateAlgorithm=GoldenSection(): the algorithm choice

reltol=sqrt(eps(T)): relative tolerance

abstol=eps(T): absolute tolerance

maxiter::Integer=1_000: maximum numer of iterations