Line Search Algorithms

For the mathematical background behind line search algorithms, this Pluto notebook details a quick overview with proof of convergence. Optini's [LineSearch] abstraction provides a simple interface to customize and experiment with the basic components: the search direction, the initial step length guess, and the method to select an appropriate step length. In particular, a few examples from Nocedal & Wright's Numerical Optimization have already been implemented:

Search direction: Steepest, Newton

Initial guess: StaticInitial, PreviousDecreaseInitial, QuadraticInitial

Step length selection method: StaticLineSearch, ExactLineSearch, BacktrackingLineSearch, InterpolationLineSearch, StrongWolfeLineSearch

Search Direction Interface

Function	Input	Output

order	YourType	order of information required, e.g. Optini.SecondOrder() for Newton direction

direction	YourType, state	the search direction

To define a custom search direction, you must implement order and direction on your type. The former tells Optini whether your method requires first-order and/or second-order information, while the latter computes the direction for line search. Optini will compute each iteration's state with the objective value, first-order and/or second-order information necessary for your algorithm. The following example, for instance, is how Optini defines the steepest descent direction internally:

import Optini: order, direction, FirstOrder

# Define singleton type for dispatch purposes
struct Steepest end

# Steepest descent requires first-order information
order(::Steepest) = FirstOrder()

# Since `Steepest` is now classified as `FirstOrder`, Optini will prepare `FirstOrderState`
# information at each iteration. We access the gradient in the field `∇f`, and simply define
# the search direction as the opposite of the gradient.
direction(::Steepest, state) = -state.∇f

Initial Step Length Interface

Function	Input	Output

reset!	YourType	reset the method, especially for initial guess dependent on previous iterations

update!	YourType, state, p, α	update the method, especially for initial guess dependent on previous iterations

YourType	state, p	the initial step length guess

A custom initial step length can be implemented by defining reset!, update!, and a functor to compute the initial guess. For example, a fixed step length guess might be implemented as:

# For simplicity we're only defining `Float64` initial here. Internally, however, Optini 
# uses the parametric struct `StaticInitial{T}`.
struct StaticInitial
    α::Float64
end

# Static initials don't hold any information that needs to be reset. If you are subtyping 
# `Optini.AbstractInitial{T}`, this is already the fallback behavior.
function reset!(::StaticInitial) end

# Some initial methods require information about the last search direction `p` and step 
# length `α`, to choose an initial for the current iteration and therefore need to be updated. 
# `StaticInitial`, however, doesn't need to update anything here. Similar to `reset!`, this 
# is the default fallback behavior for types subtyping `Optini.AbstractInitial{T}`.
function update!(::StaticInitial, state, p, α) end

# Return the initial step length guess, given the current `state` and the direction `p`
(si::StaticInitial)(state, p) = si.α

Step Length Selection Method Interface

Function	Input	Output

YourType	f, state, p, α₀	the appropriate step length

A custom step length selection scheme receives the objective function f, the current state, the search direction p, and the initial step length α₀ to produce the next step length. For instance, the following code defines a fixed line search method:

struct StaticLineSearch end

# Static line search method simply uses the initial step length
(sls::StaticLineSearch)(f, state, p, α₀) = α₀

References

Optini.LineSearch — Type

LineSearch{D, I, M} <: MultivariateAlgorithm

LineSearch optimizes the objective function by first choosing a direction, then search along this direction for an appropriate step.

Fields

direction::D: determines the search direction

initial::I: the method to guess an initial step length

method::: the method to compute the appropriate next step

Optini.GradientDescent — Type

GradientDescent{I, M}

Aliasing LineSearch with steepest descent.

Optini.NewtonDescent — Type

NewtonDescent{I, M}

Aliasing LineSearch with Newton method.

Optini.StaticInitial — Type

StaticInitial{T} <: AbstractInitial{T}

Static initial α₀ for line search methods. Newton and quasi-Newton algorithms for instance should use α₀ = 1 as the initial guess.

Optini.PreviousDecreaseInitial — Type

PreviousDecreaseInitial{T, D<:Ref} <: AbstractInitial{T}

PreviousDecreaseInitial assumes the first-order decrease is the same as the previous iteration:

$α₀ = αₖ₋₁\frac{∇f^\intercal_{k-1} p_{k-1}}{∇f^\intercal_k p_k}$

Fields

α::T: default step length for the first iteration

prev_decrease::D: save the last iteration's first-order decrease

Optini.QuadraticInitial — Type

QuadraticInitial{T} <: AbstractInitial{T}

QuadraticInitial assumes $ϕ(α₀) - ϕ(0) = fₖ - fₖ₋₁$ and uses quadratic interpolation with $ϕ(α₀)$, $ϕ(0)$, $ϕ′(0)$ to compute the initial guess.

Fields

α::T: default step length for the first iteration

prev_f::Ref{T}: save the last iteration's function value

Optini.StaticLineSearch — Type

StaticLineSearch

StaticLineSearch directly uses the initial guess as the new step length.

Optini.ExactLineSearch — Type

ExactLineSearch{A<:UnivariateAlgorithm, O1, O2}

ExactLineSearch computes the optimal step length using univariate methods. Since most univariate algorithms require repeated calculations of the objective to shrink the bracket around a local minimum, ExactLineSearch is expected to be expensive.

Fields

alg::A: the univariate algorithm

bracket_options::O1: named tuple of bracketing options

alg_options::O2: named tuple of the univariate algorithm options

Optini.BacktrackingLineSearch — Type

BacktrackingLineSearch{C, T}

BacktrackingLineSearch starts from an initial step length guess and gradually backtracks until the Armijo sufficient decrease condition is met.

Fields

c::C: the constant for the Armijo condition in the open interval (0, 1)

scale::T: the scale factor to shrink the step length

Optini.InterpolationLineSearch — Type

InterpolationLineSearch{C, T}

InterpolationLineSearch uses quadratic and cubic interpolations to compute the step length satisfying the Armijo sufficient decrease condition.

Fields

c::C: the constant for the Armijo condition in the open interval (0, 1)

ϵ::T: the minimum step length decrease from the initial step length

Optini.StrongWolfeLineSearch — Type

StrongWolfeLineSearch{C1, C2, S, T}

StrongWolfeLineSearch computes a step length that satifies both the Armijo sufficient decrease and the curvature conditions.

Fields

c₁::C1: Armijo condition constant

c₂::C2: curvature condition constant

scale::S: scale factor to expand step length bracket

α_max::T: maximum step length