Trust Region Algorithms

Optini.TrustRegion — Type

TrustRegion{H, M, D, R<:Ref, E} <: MultivariateAlgorithm

TrustRegion builds a quadratic model within a radius from the current point to approximate the objective function and uses the model to choose the next step.

Fields

hessian::H: the method to compute or approximate the Hessian

method::M: the method to compute the appropriate next step

Δ₀::D: the initial radius

Δₘₐₓ::D: the maximum radius

η::E: the threshold to control when TrustRegion will take a step

ηₛ::E: the threshold to control when TrustRegion shrinks the current radius

ηₑ::E: the threshold to control when TrustRegion expands the current radius

σₛ::S: the factor used to shrink the trust region radius

σₑ::S: the factor used to expand the trust region radius

Δ::R: a Ref containing the current radius