Univariate Optimization

Optini provides a collection of popular bracketing and root-finding algorithms including Bisection, Fibonacci, GoldenSection, and QuadraticFit. The optimize call in this case requires the lower and upper bounds of the search interval, for which Optini provides a convenient bracket function to find an initial bracket around the minimum.

References

Optini.optimize — Method

optimize(f::Function, lower::Real, upper::Real; kwargs...)

Compute the minimizer of the function f using the bracket (lower, upper).

Keywords

alg::UnivariateAlgorithm=GoldenSection(): the algorithm choice

reltol=sqrt(eps(T)): relative tolerance

abstol=eps(T): absolute tolerance

maxiter::Integer=1_000: maximum numer of iterations

Optini.bracket — Function

bracket(f::Function, x::T=0.0; kwargs...) where {T<:AbstractFloat}

Create a bracket around a local minimum from the intial point x. Return the bracket in a tuple if successful, and nothing otherwise.

Keywords

step=0.01: the step size to expand the bracket

scale=2.0: the scale factor for step size at each iteration

maxiter::Integer=100: the maximum number of iterations

Optini.Bisection — Type

Bisection <: UnivariateAlgorithm

Bisection search is a root-finding algorithm by maintainining a bracket of opposing derivative signs. At each iteration the algorithm shrinks the bracket by half until a critical point is located or the bracket is sufficiently small.

Optini.Fibonacci — Type

Fibonacci{T} <: UnivariateAlgorithm

The Fibonacci search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum. The ratio is computed from the Fibonacci sequence.

Optini.GoldenSection — Type

GoldenSection <: UnivariateAlgorithm

The golden section search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum with the golden ratio as an approximation to the Fibonacci sequence.

Optini.QuadraticFit — Type

QuadraticFit{T} <: UnivariateAlgorithm

The quadratic fit algorithm constructs a quadratic function from three points at each iteration as an approximation to the objective. The analytical solution to the quadratic model is examined as the new minimizer candidate to update the three points.