Optini

Optini.AbstractAlgorithm

Optini.BacktrackingLineSearch

Optini.BacktrackingLineSearch

Optini.Bisection

Optini.ExactLineSearch

Optini.ExactLineSearch

Optini.Fibonacci

Optini.Fibonacci

Optini.FirstOrderAlgorithm

Optini.GoldenSection

Optini.GradientDescent

Optini.GradientDescent

Optini.InterpolationLineSearch

Optini.InterpolationLineSearch

Optini.MultipleIdentity

Optini.MultipleIdentity

Optini.MultivariateAlgorithm

Optini.Newton

Optini.Newton

Optini.PreviousDecreaseInitial

Optini.PreviousDecreaseInitial

Optini.QuadraticFit

Optini.QuadraticFit

Optini.QuadraticInitial

Optini.QuadraticInitial

Optini.SecondOrderAlgorithm

Optini.Solution

Optini.StaticInitial

Optini.StaticInitial

Optini.StaticLineSearch

Optini.StaticLineSearch

Optini.StrongWolfeLineSearch

Optini.StrongWolfeLineSearch

Optini.UnivariateAlgorithm

Optini.bracket

Optini.optimize

Optini.AbstractAlgorithm — Type

AbstractAlgorithm

Abstract supertype of all optimization algorithms.

Optini.BacktrackingLineSearch — Type

struct BacktrackingLineSearch{I<:AbstractInitial, C, T} <: AbstractLineSearch

BacktrackingLineSearch starts from an initial step length guess and gradually backtracks until the Armijo sufficient decrease condition is met.

Fields

init::I: the initial step length method

c::C: the constant for the Armijo condition in the open interval (0, 1)

scale::T: the scale factor to shrink the step length

Optini.BacktrackingLineSearch — Method

BacktrackingLineSearch(; kwargs...)

Initiate BacktrackingLineSearch.

Keywords

init=StaticInitial(): the initial step length method

c=1e-4: the constant for the Armijo condition in the open interval (0, 1)

scale=0.5: the scale factor to shrink the step length

Optini.Bisection — Type

Bisection <: UnivariateAlgorithm

Bisection search is a root-finding algorithm by maintainining a bracket of opposing derivative signs. At each iteration the algorithm shrinks the bracket by half until a critical point is located or the bracket is sufficiently small.

Optini.ExactLineSearch — Type

ExactLineSearch{I<:AbstractInitial, A<:UnivariateAlgorithm, O1, O2} <: AbstractLineSearch

ExactLineSearch computes the optimal step length using univariate methods. Since most univariate algorithms require repeated calculations of the objective to shrink the bracket around a local minimum, ExactLineSearch is expected to be expensive.

Fields

init::I: the initial step length method

alg::A: the univariate algorithm

bracket_options::O1: named tuple of bracketing options

alg_options::O2: named tuple of the univariate algorithm options

Optini.ExactLineSearch — Method

ExactLineSearch(; kwargs...)

Initiate a ExactLineSearch. Use keyword arguments to tweak the bracketing phase and the univariate algorithm.

Keywords

init=StaticInitial(): the initial step length method

step=0.01: bracketing step size

scale=2.0: bracketing scale factor

bracket_max_iter::Integer=100: bracketing maximum number of iterations

alg::UnivariateAlgorithm: the univariate algorithm to optimize the step length

rel_tol=1e-6: relative tolerance

abs_tol=1e-12: absolute tolerance

max_iter::Integer=1_000: maximum number of iterations

Optini.Fibonacci — Type

Fibonacci{T} <: UnivariateAlgorithm

The Fibonacci search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum. The ratio is computed from the Fibonacci sequence.

Optini.Fibonacci — Type

Fibonacci([ϵ=0.01])

Initialize Fibonacci search algorithm with optional argument ϵ in the interval (0, 1) to control how close the last iterate is to the previous one. Refer to Kochenderfer and Wheeler's "Algorithms for Optimization" for more information.

Optini.FirstOrderAlgorithm — Type

FirstOrderAlgorithm

Abstract supertype of multivariate optimization algorithms requiring first-order knowledge.

Optini.GoldenSection — Type

GoldenSection <: UnivariateAlgorithm

The golden section search algorithm optimize a univariate function by iteratively shrinking the initial bracket around the minimum with the golden ratio as an approximation to the Fibonacci sequence.

Optini.GradientDescent — Type

GradientDescent{T<:AbstractLineSearch} <: FirstOrderAlgorithm

GradientDescent uses steepest descent and line search to minimize the objective function.

Fields

linesearch::T: line search method

Optini.GradientDescent — Method

GradientDescent(; linesearch=StaticLineSearch())

Initiate GradientDescent algorithm.

Optini.InterpolationLineSearch — Type

InterpolationLineSearch{I<:AbstractInitial, C, T}

InterpolationLineSearch uses quadratic and cubic interpolations to compute the step length satisfying the Armijo sufficient decrease condition.

Fields

init::I: the initial step length method

c::C: the constant for the Armijo condition in the open interval (0, 1)

ϵ::T: the minimum step length decrease from the initial step length

Optini.InterpolationLineSearch — Method

InterpolationLineSearch(; kwargs...)

Initiate InterpolationLineSearch.

Keywords

init=StaticInitial(): the initial step length method

c=1e-4: the constant for the Armijo condition in the open interval (0, 1)

ϵ=1e-6: the minimum step length decrease from the initial step length

Optini.MultipleIdentity — Type

MultipleIdentity{T1, T2}

Modify the Hessian by adding some multiples, τ, of the identity matrix, as described in Nocedal and Wright's Numerical Optimization.

Fields

β::T1: constant to compute appropriate τ

ρ::T2: scale factor to expand τ

Optini.MultipleIdentity — Method

MultipleIdentity(; beta=1e-3, scale=2.0)

Initiate MultipleIdentity method to modify the Hessian.

Optini.MultivariateAlgorithm — Type

MultivariateAlgorithm

Abstract supertype of multivariate optimization algorithms.

Optini.Newton — Type

Newton{LS<:AbstractLineSearch, M} <: SecondOrderAlgorithm

Newton method uses the Newton direction and line search to minimize the objective.

Fields

linesearch::LS: line search method

modify::M: the Hessian modification method to enforce positive definiteness

Optini.Newton — Method

Newton(; linesearch=StaticLineSearch(), modify=MultipleIdentity())

Initiate Newton algorithm.

Optini.PreviousDecreaseInitial — Type

PreviousDecreaseInitial{T, D<:Ref} <: AbstractInitial{T}

PreviousDecreaseInitial assumes the first-order decrease is the same as the previous iteration:

$α₀ = αₖ₋₁\frac{∇f^\intercal_{k-1} p_{k-1}}{∇f^\intercal_k p_k}$

Fields

α::T: default step length for the first iteration

prev_decrease::D: save the last iteration's first-order decrease

Optini.PreviousDecreaseInitial — Method

PreviousDecreaseInitial([α::T=1.0]) where {T}

Initialize PreviousDecreaseInitial, with prev_decrease defaulting to Ref(T(NaN)).

Optini.QuadraticFit — Type

QuadraticFit{T} <: UnivariateAlgorithm

The quadratic fit algorithm constructs a quadratic function from three points at each iteration as an approximation to the objective. The analytical solution to the quadratic model is examined as the new minimizer candidate to update the three points.

Optini.QuadraticFit — Type

QuadraticFit([ϵ=0.01])

Initialize quadratic fit search algorithm with optional argument ϵ in the interval (0, 1) to choose a midpoint as mid = ϵ * lower + (1 - ϵ) * upper.

Optini.QuadraticInitial — Type

QuadraticInitial{T} <: AbstractInitial{T}

QuadraticInitial assumes $ϕ(α₀) - ϕ(0) = fₖ - fₖ₋₁$ and uses quadratic interpolation with $ϕ(α₀)$, $ϕ(0)$, $ϕ′(0)$ to compute the initial guess.

Fields

α::T: default step length for the first iteration

prev_f::Ref{T}: save the last iteration's function value

Optini.QuadraticInitial — Method

QuadraticInitial([α::T=1.0]) where {T}

Initialize QuadraticInitial, with prev_f defaulting to Ref(T(NaN)).

Optini.SecondOrderAlgorithm — Type

SecondOrderAlgorithm

Abstract supertype of multivariate optimization algorithms requiring second-order knowledge.

Optini.Solution — Type

Solution{Tx, Tf}

Solution type for optimization models.

Optini.StaticInitial — Type

StaticInitial{T} <: AbstractInitial{T}

Static initial α₀ for line search methods. Newton and quasi-Newton algorithms for instance should use α₀ = 1 as the initial guess.

Optini.StaticInitial — Type

StaticInitial([α=1.0])

Initialize StaticInitial{T}.

Optini.StaticLineSearch — Type

StaticLineSearch{T<:AbstractInitial}

StaticLineSearch directly uses the initial guess as the new step length.

Fields

init::T: the initial step length method

Optini.StaticLineSearch — Method

StaticLineSearch(; init=StaticInitial())

Initiates StaticLineSearch, initial guess defaults to StaticInitial.

Optini.StrongWolfeLineSearch — Type

StrongWolfeLineSearch{I<:AbstractInitial, C, S, T} <: AbstractLineSearch

StrongWolfeLineSearch computes a step length that satifies both the Armijo sufficient decrease and the curvature conditions.

Fields

init::I: initial step length method

c_1::C: Armijo condition constant

c_2::C: curvature condition constant

scale::S: scale factor to expand step length bracket

α_max::T: maximum step length

Optini.StrongWolfeLineSearch — Method

StrongWolfeLineSearch(; kwargs...)

Initiate StrongWolfeLineSearch.

Keywords

init=StaticInitial(): initial step length method

c_1=1e-4: Armijo condition constant

`c_2=0.9: curvature condition constant

scale=2.0: scale factor to expand step length bracket

α_max=100_000: maximum step length

Optini.UnivariateAlgorithm — Type

UnivariateAlgorithm

Abstract supertype of univariate optimization algorithms.

Optini.bracket — Method

bracket(f::Function, x::T=0.0; kwargs...) where {T<:AbstractFloat}

Create a bracket around a local minimum from the intial point x. Return the bracket in a tuple if successful, and nothing otherwise.

Keywords

step=0.01: the step size to expand the bracket

scale=2.0: the scale factor for step size at each iteration

max_iter::Integer=100: the maximum number of iterations

Optini.optimize — Method

optimize(f::Function, lower::Real, upper::Real; kwargs...)

Compute the minimizer of the function f using the bracket (lower, upper).

Keywords

alg::UnivariateAlgorithm=GoldenSection(): the algorithm choice

rel_tol=sqrt(eps(T)): relative tolerance

abs_tol=eps(T): absolute tolerance

max_iter::Integer=1_000: maximum numer of iterations